dev/portail_django/requirements.txt
arrow==1.2.3
asgiref==3.6.0
binutils==0.2.dev0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.1
django-location-field==2.7.0
djangorestframework==3.14.0
Flask==2.2.3
folium==0.14.0
GDAL @ file:///C:/Users/mathi/Documents/MDT/TM/portail_deplacements/dev/portail_django/GDAL-3.4.3-cp311-cp311-win_amd64.whl
geos==0.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.5.0
proj==0.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3